refactor: tests/run-mypy: Extract MessageBox tests into test_messages.
This extracts MessageBox tests into their own test_messages.py file, parallel
to the previous extraction of MessageBox into their own file - messages.py.

This allows various imports and definitions to be split/moved/tidied.

An entry is added to tools/run-mypy to exclude the new file from type checking
for now, since the source file for the extracted tests, test_ui_tools.py, also
does not have type annotations.
